added recording time column in tour segmenter
git-svn-id: https://mytourbook.svn.sourceforge.net/svnroot/mytourbook/trunk@1179 5a91af82-4d22-0410-8bcb-a41d7c89129f

+++ net.tourbook/src/net/tourbook/data/TourSegment.java
@@ -40,8 +40,8 @@ public class TourSegment {
 	public float	gradient;
 	public int		power;
 
-	public float	pace;
-	public float	paceDiff;
+	public int		pace;
+	public int		paceDiff;
 	public int		pulse;
 	public int		pulseDiff;
